implemented some comments from naomi
1. The input "heritability, h2" for QT or "Trait heritability (on liability
scale, h2)"  for CC  needs to be expanded, for example add a line
underneath in grey "The power calculation requires an estimate of the
true SNP-heritability, so that the power is the probability of
estimating a SNP-heritability that is greater than zero". In fact I
think this could be moved to the options section since it is to do with
the power, and set the default as 0.20, justified by experience of
analysis of complex traits. Moving it to the option section would then
make it clear in the rg calculator that the h2 input is needed for the
estimation of the se not just power

2. "Type I error rate" update to "Type I error rate used in the power
calculation"

3. For the Variance of SNP-served genetic relationships. In the "Note:" can
you give any reason why people should change this

4. The Output Notes say "Power: Probability of detecting h2 given the type
I error rate",  replace by "Probability of detecting h2 significantly >
0 for the given the type I error rate and the inputted SNP-heritability
assumed in the population" and for rg this should be expanded further

5. In CC Replace "prevalence" by "Disease risk in population", if you don't
take up suggestion too move h2 to options then, move disease risk above
heritability.

6. For the rg calculator, it is clear that the estimates of se of rg are
highly dependent on the h2 and rg input values. I think more
instructions are needed to highlight this, since you shouldn't assume
users will actually read the paper.

7. For bivariate case-control, there is the option "are both sample used
for both traits" I don't think this makes sense for case-control when
most controls will be no for both disorders. Instead it should be made
clear that if there is a single set of controls then they should be
shared between the two disorder groups for the analysis.

Changed 1, 2, 4, 5